+++ src/test/php/net/xp_framework/unittest/util/TimeZoneTest.class.php
@@ -5,64 +5,59 @@ use util\Date;
 use util\TimeZone;
 
 class TimeZoneTest extends \unittest\TestCase {
-  private $fixture;
-
-  /** @return void */
-  public function setUp() {
-    $this->fixture= new TimeZone('Europe/Berlin');
-  }    
 
   #[@test]
   public function name() {
-    $this->assertEquals('Europe/Berlin', $this->fixture->getName());
+    $this->assertEquals('Europe/Berlin', (new TimeZone('Europe/Berlin'))->name());
   }
-  
-  #[@test]
-  public function offsetDST() {
-    $this->assertEquals('+0200', $this->fixture->getOffset(new Date('2007-08-21')));
+
+  #[@test, @values([
+  #  ['Europe/Berlin', '2007-08-21', 7200],
+  #  ['Australia/Adelaide', '2007-01-21', 37800],
+  #  ['America/New_York', '2007-08-21', -14400]
+  #])]
+  public function offset_daylight($tz, $date, $offset) {
+    $this->assertEquals($offset, (new TimeZone($tz))->offset(new Date($date)));
   }
-  
-  #[@test]
-  public function offsetNoDST() {
-    $this->assertEquals('+0100', $this->fixture->getOffset(new Date('2007-01-21')));
+
+  #[@test, @values([
+  #  ['Europe/Berlin', '2007-01-21', 3600],
+  #  ['Australia/Adelaide', '2007-08-21', 34200],
+  #  ['America/New_York', '2007-01-21', -18000]
+  #])]
+  public function offset_standard($tz, $date, $offset) {
+    $this->assertEquals($offset, (new TimeZone($tz))->offset(new Date($date)));
   }
-  
-  #[@test]
-  public function offsetWithHalfHourDST() {
-    // Australia/Adelaide is +10:30 in DST
-    $this->fixture= new TimeZone('Australia/Adelaide');
-    $this->assertEquals('+1030', $this->fixture->getOffset(new Date('2007-01-21')));
+
+  #[@test, @values([
+  #  ['Europe/Berlin', '2007-08-21', '+0200'],
+  #  ['Australia/Adelaide', '2007-01-21', '+1030'],
+  #  ['America/New_York', '2007-08-21', '-0400']
+  #])]
+  public function difference_daylight($tz, $date, $diff) {
+    $this->assertEquals($diff, (new TimeZone($tz))->difference(new Date($date)));
   }
-  
-  #[@test]
-  public function offsetWithHalfHourNoDST() {
-    // Australia/Adelaide is +09:30 in non-DST
-    $this->fixture= new TimeZone('Australia/Adelaide');
-    $this->assertEquals('+0930', $this->fixture->getOffset(new Date('2007-08-21')));
+
+  #[@test, @values([
+  #  ['Europe/Berlin', '2007-01-21', '+0100'],
+  #  ['Australia/Adelaide', '2007-08-21', '+0930'],
+  #  ['America/New_York', '2007-01-21', '-0500']
+  #])]
+  public function difference_standard($tz, $date, $diff) {
+    $this->assertEquals($diff, (new TimeZone($tz))->difference(new Date($date)));
   }
-  
+
   #[@test]
-  public function offsetInSecondsDST() {
-    $this->assertEquals(7200, $this->fixture->getOffsetInSeconds(new Date('2007-08-21')));
-  }
-  
-  #[@test]
-  public function offsetInSecondsNoDST() {
-    $this->assertEquals(3600, $this->fixture->getOffsetInSeconds(new Date('2007-01-21')));
-  }
-  
-  #[@test]
-  public function convert() {
-    $date= new Date('2007-01-01 00:00 Australia/Sydney');
+  public function translate() {
     $this->assertEquals(
       new Date('2006-12-31 14:00:00 Europe/Berlin'),
-      $this->fixture->translate($date)
+      (new TimeZone('Europe/Berlin'))->translate(new Date('2007-01-01 00:00 Australia/Sydney'))
     );
   }
   
   #[@test]
   public function previousTransition() {
-    $transition= $this->fixture->previousTransition(new Date('2007-08-23'));
+    $transition= (new TimeZone('Europe/Berlin'))->previousTransition(new Date('2007-08-23'));
     $this->assertEquals(true, $transition->isDst());
     $this->assertEquals('CEST', $transition->abbr());
     $this->assertEquals('+0200', $transition->difference());
@@ -71,7 +66,7 @@ class TimeZoneTest extends \unittest\TestCase {
   
   #[@test]
   public function previousPreviousTransition() {
-    $transition= $this->fixture->previousTransition(new Date('2007-08-23'));
+    $transition= (new TimeZone('Europe/Berlin'))->previousTransition(new Date('2007-08-23'));
     $previous= $transition->previous();
     $this->assertFalse($previous->isDst());
     $this->assertEquals('CET', $previous->abbr());
@@ -81,7 +76,7 @@ class TimeZoneTest extends \unittest\TestCase {
 
   #[@test]
   public function previousNextTransition() {
-    $transition= $this->fixture->previousTransition(new Date('2007-08-23'));
+    $transition= (new TimeZone('Europe/Berlin'))->previousTransition(new Date('2007-08-23'));
     $next= $transition->next();
     $this->assertFalse($next->isDst());
     $this->assertEquals('CET', $next->abbr());
@@ -91,7 +86,7 @@ class TimeZoneTest extends \unittest\TestCase {
 
   #[@test]
   public function nextTransition() {
-    $transition= $this->fixture->nextTransition(new Date('2007-08-23'));
+    $transition= (new TimeZone('Europe/Berlin'))->nextTransition(new Date('2007-08-23'));
     $this->assertEquals(false, $transition->isDst());
     $this->assertEquals('CET', $transition->abbr());
     $this->assertEquals('+0100', $transition->difference());
@@ -101,5 +96,49 @@ class TimeZoneTest extends \unittest\TestCase {
   #[@test, @expect(IllegalArgumentException::class)]
   public function unknownTimeZone() {
     new TimeZone('UNKNOWN');
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function name_getter() {
+    $this->assertEquals('Europe/Berlin', (new TimeZone('Europe/Berlin'))->getName());
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function offsetDST() {
+    $this->assertEquals('+0200', (new TimeZone('Europe/Berlin'))->getOffset(new Date('2007-08-21')));
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function offsetNoDST() {
+    $this->assertEquals('+0100', (new TimeZone('Europe/Berlin'))->getOffset(new Date('2007-01-21')));
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function offsetWithHalfHourDST() {
+    // Australia/Adelaide is +10:30 in DST
+    $this->assertEquals('+1030', (new TimeZone('Australia/Adelaide'))->getOffset(new Date('2007-01-21')));
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function offsetWithHalfHourNoDST() {
+    // Australia/Adelaide is +09:30 in non-DST
+    $this->assertEquals('+0930', (new TimeZone('Australia/Adelaide'))->getOffset(new Date('2007-08-21')));
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function offsetInSecondsDST() {
+    $this->assertEquals(7200, (new TimeZone('Europe/Berlin'))->getOffsetInSeconds(new Date('2007-08-21')));
+  }
+
+  /** @deprecated */
+  #[@test]
+  public function offsetInSecondsNoDST() {
+    $this->assertEquals(3600, (new TimeZone('Europe/Berlin'))->getOffsetInSeconds(new Date('2007-01-21')));
   }
 }